Best Classifier notebook corrected
It seems that after a second review it is not yet clear that
things are correct, so I tried to explain with text, and showing
some results more explicitly, like F1 score and Jaccard similarity
index, that I completely cover the requirements from the rubric.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	modified:   Best_Classifier.ipynb